+++ turkle/static/admin/js/turkle-admin.js
@@ -1,9 +1,13 @@
 $(function () {
-  /* Hide/show group permissions on create/edit forms */
+  /* On Add/Change forms, use value of 'Custom Permissions' checkbox to hide/show
+     widgets for selecting Groups and Users that should be assigned 'can_work_on'
+     permissions for the model instance being added/changed */
   if (!$('#id_custom_permissions').is(':checked')) {
-    $('div.field-worker_permissions').hide();
+    $('div.field-can_work_on_groups').hide();
+    $('div.field-can_work_on_users').hide();
   }
   $('#id_custom_permissions').change(function() {
-    $('div.field-worker_permissions').toggle();
+    $('div.field-can_work_on_groups').toggle();
+    $('div.field-can_work_on_users').toggle();
   });
 });