+++ CHANGELOG.md
@@ -30,6 +30,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   restricted to superusers
 - On the Change User admin form, the "Last login" and "Date joined"
   fields are now read-only
+- Permission to work on a Batch can now be restricted to specific Users.
+  Previously, Batch access could only be restricted at the Group level,
+  which in practice lead to the creation of many single-User Groups.
 - Reorganized Add/Change Project forms to more clearly identify which
   Project-level settings are only used as defaults for new Batches.
   Reorganized Add/Change Batch forms to match field order on Add/Change